Feature #70119, TouchAssetController service proxy (#603)
