+++ core/src/main/java/inetsoft/web/viewsheet/event/TouchAssetEvent.java
@@ -18,9 +18,11 @@
 package inetsoft.web.viewsheet.event;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import org.immutables.serial.Serial;
 import org.immutables.value.Value;
 
 @Value.Immutable
+@Serial.Structural
 @JsonDeserialize(as = ImmutableTouchAssetEvent.class)
 public abstract class TouchAssetEvent {
    public abstract boolean design();
